@@ -1,6 +1,6 @@
 {
   "name": "@wslyhbb/vuejs-datepicker",
-  "version": "3.0.3",
+  "version": "4.0.0",
   "publishConfig": {
     "access": "public"
   },
@@ -41,7 +41,7 @@
     "build:main:cjs": "rollup --config build/rollup.config.cjs.mjs",
     "prepublishOnly": "npm run lint && npm run build",
     "dev": "cross-env NODE_ENV=development rollup --config build/rollup.config.es.js --watch",
-    "serve": "rollup -c build/serve.mjs --watch",
+    "serve": "vite --mode loc",
     "test": "jest --config test/unit/jest.conf.js --coverage",
     "lint": "eslint --ext .js,.vue,.ts src test/unit/specs"
   },
@@ -53,7 +53,7 @@
     "date-fns": "^2.29.3"
   },
   "peerDependencies": {
-    "vue": "^2.6.10"
+    "vue": "^3.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.7",
@@ -65,8 +65,10 @@
     "@rollup/plugin-commonjs": "^24.0.0",
     "@rollup/plugin-node-resolve": "^15.0.1",
     "@rollup/plugin-terser": "^0.2.1",
-    "@vue/test-utils": "^1.3.3",
-    "@vue/vue2-jest": "^29.2.2",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vue/compiler-sfc": "^3.2.45",
+    "@vue/test-utils": "^2.2.6",
+    "@vue/vue3-jest": "^29.2.2",
     "autoprefixer": "^10.4.13",
     "babel-jest": "^29.3.1",
     "clean-css": "^5.3.1",
@@ -87,9 +89,10 @@
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-replace": "^2.2.0",
     "rollup-plugin-serve": "^2.0.2",
-    "rollup-plugin-vue": "^5.1.9",
-    "vue": "^2.7.14",
-    "vue-eslint-parser": "^9.1.0",
-    "vue-template-compiler": "^2.7.14"
+    "rollup-plugin-vue": "^6.0.0",
+    "stylus": "^0.59.0",
+    "vite": "^4.0.3",
+    "vue": "^3.2.45",
+    "vue-eslint-parser": "^9.1.0"
   }
 }
